Fixed the bug, merged CancelCommand into ExecutionControlCommands
